requirements.txt
inspect-ai
inspect-wandb[weave]
inspect_evals
wandb
openai
huggingface-hub
transformers
torch
pyyaml
python-dotenv